register screen -> reg screen and pkg profile
Split registergui's 'register' async task into
'register' and 'update_package_profile', and
split PerformRegisterScreen into that and
a screen for update_package_profile.

Amongst other things this separates the error
handling blocks for the register request and the
package profile update request. This way if the
package profile upload fails, it doesn't cause
the gui to indicates register failure.It also
provides an opporturnity to indicate we are doing a
package profile upload in the gui.